Add color component
